In this update, several changes were made. Firstly, two directories 'docs/' and 'tmp/' were added to the .gitignore file, which means these folders will be ignored by git and changes in them will not be tracked.
The README.md file was updated to change the link to the XML-based configuration guide. It now points to a markdown file within the 'docs' directory.

In the 'DynamoDBEntityWithHashAndRangeKeyCriteria.java' file, a method was created to handle attribute value conversion. This improves code readability and maintainability by reducing redundancy in the code.

A new file named '_config.yml' was added in the 'docs' directory. This file is used to configure the settings for a Jekyll site, a simple, blog-aware static site generator. This suggests that the project is setting up a Jekyll site for documentation purposes. The configuration file contains site settings such as title, email, description, base URL, twitter username, and Github username. It also includes build settings like the theme and plugins.

Overall, from a business perspective, the project is improving its code structure and setting up a documentation site for better readability and accessibility for users or developers.

Signed-off-by: Nader Fares <naderfares@gmail.com>